+++ src/components/proofs/List.tsx
@@ -1,24 +1,21 @@
-import { useEffect } from 'preact/hooks'
-import EmailDomainStore from 'stores/EmailDomainStore'
+import { CategoriesTitles, categories } from 'models/Categories'
+import { useParams } from 'react-router-dom'
 import ListTitle from 'components/proofs/ListTitle'
 import ProofList from 'components/proofs/ProofList'
-import useUrlParams from 'hooks/useUrlParams'
 
 export default function () {
-  const params = useUrlParams()
-  const currentCategory = !!params && params?.token ? 'Email' : 'NFTs'
-
-  useEffect(() => {
-    if (params) {
-      EmailDomainStore.emailDomain = params.domain
-      EmailDomainStore.hasToken = !!params
-    }
-  }, [params])
+  const params = useParams()
+  const urlCategory = params?.category
+  const currentCategory =
+    urlCategory &&
+    Object.keys(categories).find((category) => category === urlCategory)
+      ? urlCategory
+      : Object.keys(categories)[0]
 
   return (
     <>
       <ListTitle />
-      <ProofList selectedCategory={currentCategory} />
+      <ProofList selectedCategory={currentCategory as CategoriesTitles} />
     </>
   )
 }
