+++ src/models/Categories.tsx
@@ -15,6 +15,7 @@ export interface CategoriesComponentProps {
 
 interface Category {
   [title: string]: {
+    title: string
     icon: JSX.Element
     contentToRender: JSX.Element | null
     disabled?: boolean
@@ -22,7 +23,8 @@ interface Category {
 }
 
 export const categories: Category = {
-  NFTs: {
+  nfts: {
+    title: 'NFTs',
     icon: <Nft inheritStrokeColor />,
     contentToRender: (
       <>
@@ -37,7 +39,8 @@ export const categories: Category = {
       </>
     ),
   },
-  Email: {
+  email: {
+    title: 'Email',
     icon: (
       <div className={width('w-5')}>
         <Email inheritStrokeColor />
@@ -46,6 +49,7 @@ export const categories: Category = {
     contentToRender: <EmailProofSection dataKey="Email" />,
   },
   'Assets (coming soon)': {
+    title: 'Assets (coming soon)',
     icon: <Coin inheritStrokeColor />,
     contentToRender: null,
     disabled: true,
